+++ src/lib.rs
@@ -59,6 +59,7 @@ pub mod types;
 #[cfg(not(target_arch = "wasm32"))]
 pub use net::dns::{DnsError, DnsResolver, Resolving};
 pub use net::events::NetEvent;
+pub use net::fetch_metadata::{RequestDestination, RequestMode, SecFetchSite};
 pub use net::fetcher::{Fetcher, FetcherConfig};
 pub use net::fetcher_context::{FetcherContext, NullContext};
 #[cfg(not(target_arch = "wasm32"))]